requirements.txt
# Smart factory pipeline (see CONFIG_KEYS.md)
ultralytics>=8.2.0
opencv-python>=4.8.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0.0
jinja2>=3.1.0
rembg[cpu]>=2.0.0
pillow>=10.0.0
imageio-ffmpeg>=0.4.0
torchvision>=0.16.0
# torch: install per https://pytorch.org/ (CUDA vs CPU)
